requirements.txt
Django==2.1.2
django-adminlte2==0.4.1
djangorestframework==3.9.0
lml==0.0.4
odfpy==1.3.5
pyexcel-io==0.5.9.1
pyexcel-ods==0.5.3
pytz==2018.7
Unidecode==1.0.22
